test(account): move stateless regression test into account.test.ts
Integrate the stateless account-resolution test alongside the existing
account/getAccessToken tests instead of a separate file.
