+++ packages/better-auth/src/api/routes/account.test.ts
@@ -8,13 +8,16 @@ import {
 	afterEach,
 	assert,
 	beforeAll,
+	beforeEach,
 	describe,
 	expect,
 	it,
 	vi,
 } from "vitest";
+import { betterAuth } from "../../auth/minimal";
 import { parseSetCookieHeader } from "../../cookies";
 import { signJWT, symmetricDecodeJWT } from "../../crypto";
+import { genericOAuth } from "../../plugins/generic-oauth";
 import { getTestInstance } from "../../test-utils/test-instance";
 import type { Account } from "../../types";
 import { DEFAULT_SECRET } from "../../utils/constants";
@@ -2037,5 +2040,140 @@ describe("token routes cookie cache revocation", async () => {
 			},
 		);
 		expect(bypass.error?.status).toBe(401);
+	});
+});
+
+describe("account resolution in stateless mode", async () => {
+	const IDP = "https://idp.stateless.test";
+	const STATELESS_SECRET = "stateless-test-secret-stateless-test-secret";
+
+	const idpHandlers = [
+		http.get(`${IDP}/.well-known/openid-configuration`, () =>
+			HttpResponse.json({
+				issuer: IDP,
+				authorization_endpoint: `${IDP}/authorize`,
+				token_endpoint: `${IDP}/token`,
+				userinfo_endpoint: `${IDP}/userinfo`,
+				jwks_uri: `${IDP}/jwks`,
+			}),
+		),
+		http.post(`${IDP}/token`, () =>
+			HttpResponse.json({
+				token_type: "Bearer",
+				access_token: "idp-access-token",
+				refresh_token: "idp-refresh-token",
+				expires_in: 3600,
+				scope: "openid profile email",
+			}),
+		),
+	];
+
+	beforeEach(() => server.use(...idpHandlers));
+
+	const makeStatelessAuth = () =>
+		betterAuth({
+			secret: STATELESS_SECRET,
+			baseURL: "http://localhost:3000",
+			trustedOrigins: ["http://localhost:3000"],
+			// No database -> stateless: account and session live in cookies only.
+			session: {
+				cookieCache: {
+					enabled: true,
+					maxAge: 60 * 60 * 24,
+					refreshCache: true,
+				},
+			},
+			account: { storeStateStrategy: "cookie", storeAccountCookie: true },
+			plugins: [
+				genericOAuth({
+					config: [
+						{
+							providerId: "idp",
+							clientId: "client-id",
+							clientSecret: "client-secret",
+							scopes: ["openid", "profile", "email"],
+							discoveryUrl: `${IDP}/.well-known/openid-configuration`,
+							getUserInfo: async () => ({
+								id: "shared-idp-user",
+								email: "user@stateless.test",
+								name: "Stateless User",
+								emailVerified: true,
+							}),
+						},
+					],
+				}),
+			],
+		});
+
+	type Jar = Map<string, string>;
+	const collect = (res: Response, jar: Jar) => {
+		for (const cookie of res.headers.getSetCookie()) {
+			const [pair] = cookie.split(";");
+			const idx = pair.indexOf("=");
+			jar.set(pair.slice(0, idx), pair.slice(idx + 1));
+		}
+	};
+	const cookieHeader = (jar: Jar) =>
+		[...jar].map(([name, value]) => `${name}=${value}`).join("; ");
+
+	// Sign the shared IdP user in on a fresh stateless instance, returning its cookie jar.
+	const signIn = async (auth: ReturnType<typeof makeStatelessAuth>) => {
+		const jar: Jar = new Map();
+		let res = await auth.handler(
+			new Request("http://localhost:3000/api/auth/sign-in/oauth2", {
+				method: "POST",
+				headers: { "content-type": "application/json" },
+				body: JSON.stringify({ providerId: "idp", callbackURL: "/" }),
+			}),
+		);
+		collect(res, jar);
+		const { url } = (await res.json()) as { url: string };
+		const state = new URL(url).searchParams.get("state");
+		assert(state, "expected an OAuth state to be issued");
+		res = await auth.handler(
+			new Request(
+				`http://localhost:3000/api/auth/oauth2/callback/idp?code=test-code&state=${state}`,
+				{ headers: { cookie: cookieHeader(jar) }, redirect: "manual" },
+			),
+		);
+		collect(res, jar);
+		const session = (await auth.api.getSession({
+			headers: new Headers({ cookie: cookieHeader(jar) }),
+		})) as { user: { id: string } } | null;
+		return { jar, userId: session?.user.id };
+	};
+
+	// On serverless, separate stateless instances mint different ids for the same IdP user, so
+	// the session cookie and the account cookie can end up carrying different user ids.
+	const signInOnTwoInstances = async () => {
+		const a = await signIn(makeStatelessAuth());
+		const b = await signIn(makeStatelessAuth());
+		assert(a.userId && b.userId);
+		expect(a.userId).not.toBe(b.userId);
+
+		const accountCookieName = (await makeStatelessAuth().$context).authCookies
+			.accountData.name;
+		// Browser keeps the session cookie from A but the account cookie was rewritten by B.
+		const mixed: Jar = new Map(a.jar);
+		for (const key of [...mixed.keys()]) {
+			if (key.startsWith(accountCookieName)) mixed.delete(key);
+		}
+		for (const [key, value] of b.jar) {
+			if (key.startsWith(accountCookieName)) mixed.set(key, value);
+		}
+		return { mixed, sessionUserId: a.userId, accountCookieName };
+	};
+
+	it("resolves a valid account cookie whose userId differs from the session user", async () => {
+		const { mixed, sessionUserId } = await signInOnTwoInstances();
+
+		// A fresh instance has an empty in-memory store, so resolution depends entirely on the
+		// account cookie (mirrors a serverless instance that didn't handle the sign-in).
+		const result = await makeStatelessAuth().api.getAccessToken({
+			body: { providerId: "idp", userId: sessionUserId },
+			headers: new Headers({ cookie: cookieHeader(mixed) }),
+		});
+
+		expect(result.accessToken).toBe("idp-access-token");
 	});
 });
